pybabilib/pyproject.toml
[tool.poetry]
name = "babilib"
version = "0.1.0"
description = "DAQ scripts for MPV configuration"
authors = ["KodaiOkawa <okawa@cns.s.u-tokyo.ac.jp>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sockets = "^1.0.0"
pyyaml = "^6.0.1"
pytest = "^8.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
